requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
google-generativeai
python-dotenv